+++ ui/__init__.py
@@ -1,4 +1,6 @@
-from ui.main_window import MainWindow
-from ui.i18n import i18n
+"""UI пакет."""
+from .main_window import MainWindow
+from .i18n import i18n
+from .file_selection_dialog import FileSelectionDialog
 
-__all__ = ['MainWindow', 'i18n']
+__all__ = ['MainWindow', 'i18n', 'FileSelectionDialog']